move files from events to server (correct location)
