+++ src/routes/jira/jira-router.ts
@@ -3,9 +3,9 @@ import { JiraConfigurationRouter } from "./configuration/jira-configuration-rout
 import { JiraSyncPost } from "./jira-sync-post";
 import { JiraAtlassianConnectGet } from "./jira-atlassian-connect-get";
 import { JiraEventsRouter } from "./events/jira-events-router";
-import { JiraSelectVersionRouter } from "./events/jira-select-version-router";
+import { JiraSelectVersionRouter } from "./server/jira-select-version-router";
 import { JiraContextJwtTokenMiddleware } from "middleware/jira-jwt-middleware";
-import { JiraServerUrlRouter } from "./events/jira-server-url-router";
+import { JiraServerUrlRouter } from "./server/jira-server-url-router";
 
 export const JiraRouter = Router();
