+++ src/routes/jira/server/jira-select-version-router.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { JiraJwtTokenMiddleware } from "middleware/jira-jwt-middleware";
-import { JiraSelectVersionGet } from "../server/jira-select-version-get";
+import { JiraSelectVersionGet } from "./jira-select-version-get";
 import { csrfMiddleware } from "middleware/csrf-middleware";
 
 export const JiraSelectVersionRouter = Router();
