+++ src/routes/jira/server/jira-server-url-router.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { JiraJwtTokenMiddleware } from "middleware/jira-jwt-middleware";
-import { JiraServerUrlGet } from "../server/jira-server-url-get";
+import { JiraServerUrlGet } from "./jira-server-url-get";
 import { csrfMiddleware } from "middleware/csrf-middleware";
 
 export const JiraServerUrlRouter = Router();
